Can communicate with both DG822 and SDS1072CML, send a SINE and plot the acquired waveform.
Updated frequency sweep so that it calculates the correct waiting times between trigger and waveform transfer. Still a pain. Wish it could be made better.
Added waveform to manage data

Updated frequency sweep measurement so that it would not skip measurements.
Updated FFT
Added a feq queries to scope and siggen

Added electrools with useful functions related to EE

Fixed the immense delays in the waveform acquisition by changing the status querying protocol.